add types to function arguments except for the rings/fields K

+++ src/prime_range.jl
@@ -1,4 +1,4 @@
-function prime_range(n)
+function prime_range(n::Int)
   primes = fill(true, n)
   if n == 0
     return primes
